refactor task runner
