+++ test/project_types/script/layers/infrastructure/languages/typescript_task_runner_test.rb
@@ -4,9 +4,9 @@ require "project_types/script/test_helper"
 describe Script::Layers::Infrastructure::Languages::TypeScriptTaskRunner do
   include TestHelpers::FakeFS
 
-  BELOW_NODE_VERSION = "v14.4.0"
-  EXACT_NODE_VERSION = "v14.5.0"
-  ABOVE_NODE_VERSION = "v14.6.0"
+  BELOW_NODE_VERSION = "v14.14.0"
+  EXACT_NODE_VERSION = "v14.15.0"
+  ABOVE_NODE_VERSION = "v14.16.0"
 
   ABOVE_NPM_VERSION = "5.2.1"
   EXACT_NPM_VERSION = "5.2.0"
